Separate inner-shell from outer shell in tests
Split the inner-shells into separate files from the outer shells for:

- test_geometry_generic::401-405,
- test_valid::inner_shell

Because the native POLY inner-shells are not handled by val3dity, but
 for testing the inner shells need to be passed separately with the
--ishell flag.

Skip the inner_pyramid tests for now, because they are replicated by
the new inner_shell tests, but some of the inner_pyramid data breaks
val3dity. They stay until the issue is resolved.