Flatten the json report and fix a bug identified by @Lkeurentjes
The behaviour now is that each Primitive get an id (order in the list if
CityJSON, 0-based), and then the "errors" is a flat list.

The "id" of each sub-primitive (no id possible with CityJSON, just order
in the arrays) is the id of a subprimitive. For instance, if we have a
MultiSolid with 2 Solids, then if an error e102 arise for one surface,
the id this surface could be something like `"2 | 0 | 37"`, which would
mean it's the 3rd Solid, the first Shell, and the 38th surface.